+++ sample/composeApp/src/wasmJsMain/kotlin/com/ae/log/sample/Main.kt
@@ -1,0 +1,13 @@
+package com.ae.log.sample
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+import kotlinx.browser.document
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    SampleState.initialize()
+    ComposeViewport(document.body!!) {
+        App()
+    }
+}